classapp/requirements.txt
torch
transformers
Pillow
Flask
shapely
numpy
Pillow
requests
oauthlib
requests_oauthlib
folium
matplotlib
beautifulsoup4
sentinelhub
pytz
flask-cors